+++ src/main/kotlin/com/eitanliu/dart/mappable/ui/JsonInputDialog.kt
@@ -10,9 +10,8 @@ import com.eitanliu.dart.mappable.extensions.createPropertyGraph
 import com.eitanliu.dart.mappable.extensions.propertyRef
 import com.eitanliu.dart.mappable.extensions.value
 import com.eitanliu.dart.mappable.generator.buildDartGenerator
-import com.eitanliu.dart.mappable.utils.ApplicationUtils
 import com.eitanliu.dart.mappable.observable.PropertyGraphWrapper
-import com.eitanliu.dart.mappable.settings.Settings
+import com.eitanliu.dart.mappable.utils.ApplicationUtils
 import com.google.gson.*
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.observable.properties.PropertyGraph